scout/requirements.txt
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
distro==1.9.0
ecdsa==0.19.1
fastapi==0.116.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.11.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.108.0
passlib==1.7.4
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
Telethon==1.41.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
